SE-612: More expressive error message for invalid column types
